requirements.txt
joblib==0.13.2
numpy==1.17.2
Pillow==6.1.0
scikit-learn==0.21.3
scipy==1.3.1
sklearn==0.0
